Move tooltip component to app/ui
