--- app/tooltip/client/index.js
@@ -1,4 +1,0 @@
-import './rocketchat-tooltip.html';
-import { tooltip } from './rocketchat-tooltip';
-
-export { tooltip };
