--- app/tooltip/index.js
@@ -1,1 +1,0 @@
-export * from './client/index';
