+++ app/ui/client/components/tooltip.html
@@ -1,4 +1,4 @@
-<template name="rocketchatTooltip">
+<template name="tooltip">
 	<div class="tooltip">
 		<div class="content">
 		</div>
